Move disk actions to a separate component
