+++ src/components/vm/disks/vmDisksCard.jsx
@@ -24,12 +24,9 @@ import { useDialogs } from 'dialogs.jsx';
 
 import { convertToUnit, diskPropertyChanged, toReadableNumber, units, vmId } from "../../../helpers.js";
 import { AddDiskModalBody } from './diskAdd.jsx';
-import { domainDetachDisk, domainGet } from '../../../libvirtApi/domain.js';
-import { EditDiskAction } from './diskEdit.jsx';
 import WarningInactive from '../../common/warningInactive.jsx';
 import { ListingTable } from "cockpit-components-table.jsx";
-import { DeleteResourceButton } from '../../common/deleteResource.jsx';
-import { DISK_SOURCE_LIST, DiskSourceCell, DiskExtras, getDiskSourceValue } from './vmDiskColumns.jsx';
+import { DiskSourceCell, DiskExtras, DiskActions } from './vmDiskColumns.jsx';
 
 const _ = cockpit.gettext;
 
@@ -233,40 +230,12 @@ export const VmDisksCard = ({ vm, disks, renderCapacity, supportedDiskBusTypes }
             });
         }
 
-        const onRemoveDisk = () => {
-            return domainDetachDisk({ connectionName: vm.connectionName, id: vm.id, name: vm.name, target: disk.target, live: vm.state === 'running', persistent: vm.persistent })
-                    .then(() => domainGet({ connectionName: vm.connectionName, id:vm.id }));
-        };
-
-        const deleteDialogProps = {
-            title: _("Remove disk from VM?"),
-            actionName: _("Remove"),
-            errorMessage: cockpit.format(_("Disk $0 could not be removed"), disk.target),
-            actionDescription: cockpit.format(_("This disk will be removed from $0:"), vm.name),
-            objectDescription: [
-                { name: _("Target"), value: <span className="ct-monospace">{disk.target}</span> },
-                ...DISK_SOURCE_LIST.flatMap(entry => getDiskSourceValue(disk.source, entry.name)
-                    ? { name: entry.label, value: <span className="ct-monospace">{getDiskSourceValue(disk.source, entry.name)}</span> }
-                    : [])
-            ],
-            deleteHandler: () => onRemoveDisk(),
-        };
-
-        const diskActions = (
-            <div className='machines-listing-actions'>
-                <DeleteResourceButton objectId={vm.name + "-disk-" + disk.target}
-                                      disabled={vm.state != 'shut off' && vm.state != 'running'}
-                                      dialogProps={deleteDialogProps}
-                                      overlayText={_("The VM needs to be running or shut off to detach this device")}
-                                      actionName={_("Remove")} />
-                { vm.persistent && vm.inactiveXML.disks[disk.target] && // supported only  for persistent disks
-                <EditDiskAction disk={disk}
-                                vm={vm}
-                                idPrefix={`${idPrefixRow}-edit`}
-                                supportedDiskBusTypes={supportedDiskBusTypes} />}
-            </div>
-        );
-        columns.push({ title: diskActions });
+        columns.push({
+            title: <DiskActions vm={vm}
+                                disk={disk}
+                                supportedDiskBusTypes={supportedDiskBusTypes}
+                                idPrefixRow={idPrefixRow} />
+        });
         return { columns, props: { key: idPrefixRow } };
     });
